+++ CHANGELOG.md
@@ -17,6 +17,11 @@ Maybe:
 - mypy passing
 - pass force-regen flag from pytest
 
+## 3.7.5
+
+- add `add_tapers_cross_section` to taper component cross_sections
+- letter `v` in text_rectangular is now DRC free
+
 ## 3.7.4
 
 - add pad_gsg_short and pad_gsg_open to components